Adds support for source maps using google-closure.
The support from google-closure is not used in a default implementation of particular Source called SourceMapNut. This interface allows to manipulate the Source object as a Nut, meaning it can be served to the client. In that case, the nut will corresponds to the source map itself and will refer all the original nuts. Source map aggregation is also supported since the google-closure allows to merge any source mapping. This commit introduces a lot of changes in the internal transformer chaining mechanism, especially to make sure that the composite information will be available when building the source map.
 Fixes #168.