Makes sure relative path is referenced in source map file.
If sources have a reachable parent file, they were not copied to the directory where source map is generated.
This was done in order to improve performances.
However, because the source and working directories are not the same, the source files were not referenced by the source map in a way they can be resolved easily.
Consequently, the approach has been changed. Now WUIC tries to use the source directory as working directory.
A configuration point has been introduced in order to avoid this behavior if needed.
In that case, all sources are copied to a temporary directory used as working directory.
This allows to give an opportunity to not copy the source files (keeping performances) and to always generate files in the directory where sources are located.

Fixes https://github.com/wuic/wuic/issues/218